requirements.txt
django
celery
requests
django-filter
djangorestframework
drf-yasg
djangorestframework-xml
djangorestframework-csv
djangorestframework-simplejwt

# dev
django-debug-toolbar
django-extensions
ipython

# flake8
flake8
flake8-print
flake8-builtins

# pytest
pytest
pytest-django
pytest-cov
pytest-mock
